refactor: use path aliases in changed files for #977 fix
Converts all changed test files to use @/* and @test/* path aliases
instead of deep relative imports. Adds missing barrel re-exports
(ParseValidationError, makeParseRetryStrategy → @/agents; validateLLMShape
→ @/review; ContextOrchestrator → @/context; executionStage/_executionDeps
→ @/pipeline) to satisfy the check:alias-internals gate. Updates the
deep-relatives baseline from 3426 → 3319 (−107).
