+++ src/pipeline/index.ts
@@ -17,3 +17,4 @@ export type { PipelineRunResult } from "./runner";
 
 export { PipelineEventEmitter } from "./events";
 export type { PipelineEvents, RunSummary } from "./events";
+export { executionStage, _executionDeps } from "./stages";
