+++ src/review/index.ts
@@ -10,3 +10,4 @@ export * from "./diff-utils";
 export * from "./finding-projection";
 export * from "./types";
 export * from "./runner";
+export { validateLLMShape } from "./semantic-helpers";
