+++ test/unit/operations/run-operation-retry-decisions.test.ts
@@ -1,11 +1,11 @@
 import { afterEach, describe, expect, test } from "bun:test";
-import type { AgentRunRequest } from "../../../src/agents";
-import { pickSelector } from "../../../src/config";
-import type { DEFAULT_CONFIG } from "../../../src/config";
-import { callOp } from "../../../src/operations/call";
-import type { RunOperation } from "../../../src/operations/types";
-import { makeMockAgentManager, makeSessionManager, makeTestRuntime } from "../../helpers";
-import type { NaxRuntime } from "../../../src/runtime";
+import type { AgentRunRequest } from "@/agents";
+import { pickSelector } from "@/config";
+import type { DEFAULT_CONFIG } from "@/config";
+import { callOp } from "@/operations";
+import type { RunOperation } from "@/operations/types";
+import { makeMockAgentManager, makeSessionManager, makeTestRuntime } from "@test/helpers";
+import type { NaxRuntime } from "@/runtime";
 
 let runtime: NaxRuntime | undefined;
 afterEach(async () => { await runtime?.close(); });
